Revert "fixed the types, and moved the reply section to a new component"
