+++ prisma/schema.prisma
@@ -51,7 +51,7 @@ model Comment {
   userEmail String
   post      Post      @relation(fields: [postId], references: [id], onDelete: Cascade)
   postId    Int
-  parent    Comment?  @relation("CommentReplies", fields: [parentId], references: [id])
-  parentId  Int?
+  reply     Comment?  @relation("CommentReplies", fields: [commentId], references: [id])
+  commentId Int?
   replies   Comment[] @relation("CommentReplies")
 }
