+++ src/app/components/PostCard.tsx
@@ -11,18 +11,19 @@ import {
   ShareIcon,
 } from "@heroicons/react/24/outline";
 import { useState } from "react";
-import { createComment } from "../server/comment-queries/queries";
+import {
+  createComment,
+  deleteCommentById,
+} from "../server/comment-queries/queries";
 import { useSession } from "next-auth/react";
 import { createLike, dislikePostPerId } from "../server/like-queries/queries";
 import { HandThumbDownIcon } from "@heroicons/react/24/outline";
-import ReplySection from "./ReplySection";
-
-type FullComment = Comment & { replies: Array<Comment>; user: User };
 
 type PostCardProps = {
   post: Post & {
     author?: User;
-    comments?: Array<FullComment & { replies: Array<FullComment> }>;
+    // comment?: Comment;
+    comments?: Array<Comment & { user: User }>;
     likes?: Array<Like & { user: User }>;
   };
 };
@@ -33,8 +34,6 @@ export function PostCard({ post }: PostCardProps) {
 
   const { data: session } = useSession();
 
-  const mainComments = post.comments?.filter((item) => !item.parentId);
-
   const hasUserLiked = !!post.likes?.find(
     (item) => item.userEmail === session?.user?.email,
   );
@@ -42,6 +41,7 @@ export function PostCard({ post }: PostCardProps) {
   const handleShare = async (payload: SharePostPayload) => {
     await sharePostById(payload);
   };
+
   const handleAddComment = async (e: React.FormEvent) => {
     e.preventDefault();
     await createComment({
@@ -65,6 +65,17 @@ export function PostCard({ post }: PostCardProps) {
           userEmail: session?.user?.email!,
         });
       }
+    } catch (error) {
+      console.log(error);
+    }
+  };
+
+  const handleDeleteComment = async (commentId: number) => {
+    try {
+      await deleteCommentById({
+        commentId: commentId,
+        userEmail: session?.user?.email!,
+      });
     } catch (error) {
       console.log(error);
     }
@@ -145,8 +156,29 @@ export function PostCard({ post }: PostCardProps) {
         </div>
         {showComments && (
           <div className="flex flex-col space-y-4">
-            {mainComments?.map((comment) => (
-              <ReplySection comment={comment} postId={post.id} />
+            {post.comments?.map((comment) => (
+              <div key={comment.id} className="flex items-center space-x-2">
+                <img
+                  className="size-8 rounded-full"
+                  src={comment.user.picture}
+                />
+                <div className="flex h-fit max-w-96 flex-col rounded-lg bg-gray-200 p-2">
+                  <p className="text-sm font-semibold">
+                    {comment.user?.firstname}
+                  </p>
+                  <p className="max-w-fit break-words" key={comment?.id!}>
+                    {comment?.content}
+                  </p>
+                  <button>
+                    <p
+                      onClick={() => handleDeleteComment(comment.id)}
+                      className="text-right text-xs text-gray-400"
+                    >
+                      Delete
+                    </p>
+                  </button>
+                </div>
+              </div>
             ))}
             <form
               className="flex justify-center space-x-2 pl-10"
