+++ src/app/server/comment-queries/queries.ts
@@ -7,7 +7,11 @@ import { CreateCommentPayload, DeleteCommentPayload } from "./types";
 export const createComment = async (payload: CreateCommentPayload) => {
   try {
     await prisma.comment.create({
-      data: payload,
+      data: {
+        content: payload.content,
+        postId: payload.postId,
+        userEmail: payload.userEmail,
+      },
     });
     revalidatePath("/posts");
   } catch (error) {
