+++ src/app/server/comment-queries/types.ts
@@ -2,7 +2,6 @@ export type CreateCommentPayload = {
   content: string;
   userEmail: string;
   postId: number;
-  parentId?: number;
 };
 
 export type DeleteCommentPayload = {
