+++ src/app/server/post-queries/queries.ts
@@ -26,7 +26,6 @@ export async function getPosts() {
       comments: {
         include: {
           user: true,
-          replies: true,
         },
       },
       likes: true,
