+++ src/app/server/reply-queries/queries.ts
@@ -1,3 +1,10 @@
 import prisma from "@/app/lib/prisma";
+import { CreateReplyOnACommentPayload } from "./types";
 
-export const createReplyOnAComment = async () => {};
+export const createReplyOnAComment = async (
+  payload: CreateReplyOnACommentPayload,
+) => {
+  await prisma.comment.create({
+    data: payload,
+  });
+};
