+++ src/app/server/reply-queries/types.ts
@@ -1,1 +1,6 @@
-export type CreateReplyOnACommentPayload = {};
+export type CreateReplyOnACommentPayload = {
+  userEmail: string;
+  commentId: number;
+  content: string;
+  postId: number;
+};
